requirements.txt
numpy
matplotlib
tensorflow
xgboost
joblib
tensorflowjs
pillow
fastapi uvicorn
scikit-learn
opencv-python
python-multipart
